fix(io): pass dtype=float32 to soundfile.read in load_with_soundfile (#3748)
`load_with_soundfile` was the only soundfile-backed loader path that
returned float64 for PCM ≥ 16-bit sources, while its sibling
`loader.py:66` already passed `dtype=np.float32`. The canonical
`unified_loader.load_audio` and the FFmpeg fallback both go through
this function, so a 7200 s × 96 kHz stereo load peaked at ~10.6 GB
instead of ~5.3 GB before `validate_audio`'s downstream cast clamped
it back. The doubled intermediate is gone now.

Kept `always_2d=False` so the mono/stereo ndim checks downstream
(\`if audio_data.ndim == 1: ... elif audio_data.ndim == 2: ...\`) keep
working unchanged — this is the minimum-change fix.

67 of the io / loader / unified_loader tests pass with no new
failures.

+++ auralis/io/loaders/soundfile_loader.py
@@ -58,8 +58,16 @@ def load_with_soundfile(file_path: Path) -> tuple[np.ndarray, int]:
                     f"({completeness:.1f}% complete)"
                 )
 
-        # Load audio data
-        audio_data, sample_rate = sf.read(str(file_path), always_2d=False)
+        # #3748: explicit float32 to match `loader.py`'s sibling sf.read
+        # call. Without the dtype hint, soundfile returns float64 for
+        # PCM ≥ 16-bit sources, which doubles peak RAM on the canonical
+        # `unified_loader.load_audio` path (a 7200s × 96 kHz stereo load
+        # goes 5.3 GB → 10.6 GB) before validate_audio's downstream cast
+        # clamps it back to float32. `always_2d=False` is preserved so
+        # the existing mono/stereo ndim checks below keep working.
+        audio_data, sample_rate = sf.read(
+            str(file_path), dtype="float32", always_2d=False
+        )
 
         # Ensure proper shape
         if audio_data.ndim == 1:
